+++ cmd/cardslurp/internal/cardfileutil/cardFileUtil_test.go
@@ -12,7 +12,7 @@ func TestIsFileSame(t *testing.T) {
 	// Torture test for boundary conditions.  :-)
 	for transBuff := 1; transBuff <= maxTransBuff; transBuff++ {
 
-		sameStat, err := IsFileSame("testData/same_a.txt", "testData/same_b.txt", transBuff)
+		sameStat, err := IsFileSame("testData/same_a.txt", "testData/same_b.txt", uint64(transBuff))
 		if err != nil {
 			fmt.Print("Error calling IsFileSame: " + err.Error() + "\n")
 			t.Fail()
@@ -22,7 +22,7 @@ func TestIsFileSame(t *testing.T) {
 			t.Error("Files were the same, but they tested as different.\n")
 		}
 
-		diffStat, err := IsFileSame("testData/same_a.txt", "testData/diff_b.txt", transBuff)
+		diffStat, err := IsFileSame("testData/same_a.txt", "testData/diff_b.txt", uint64(transBuff))
 		if err != nil {
 			fmt.Print("Error calling IsFileSame: " + err.Error() + "\n")
 		}
@@ -39,7 +39,7 @@ func TestNibbleCopy(t *testing.T) {
 
 	for transBuff := 1; transBuff <= maxTransBuff; transBuff++ {
 
-		nibStat, err := NibbleCopy("testData/same_a.txt", "testData/victim.txt", transBuff)
+		nibStat, err := NibbleCopy("testData/same_a.txt", "testData/victim.txt", uint64(transBuff))
 		if err != nil {
 			print("Error calling NibbleCopy: " + err.Error() + "\n")
 			t.Fail()
@@ -49,7 +49,7 @@ func TestNibbleCopy(t *testing.T) {
 			t.Errorf("Nibble copy returned false: transBuff == %d\n", transBuff)
 		}
 
-		sameStat, err := IsFileSame("testData/same_a.txt", "testData/victim.txt", transBuff)
+		sameStat, err := IsFileSame("testData/same_a.txt", "testData/victim.txt", uint64(transBuff))
 		if err != nil {
 			print("Error calling IsFileSame: " + err.Error() + "\n")
 			t.Fail()
